requirements.txt
gym
ALE
gym[atari]
gym[accept-rom-license]
pytorch-functional